Time: 266 ms (45.97%), Space: 123.4 MB (33.97%) - LeetHub

+++ 0234-palindrome-linked-list/0234-palindrome-linked-list.cpp
@@ -11,7 +11,7 @@
 class Solution {
 public:
     bool isPalindrome(ListNode* head) {
-        vector <int> v;
+        vector <short> v;
         while(head != NULL)
         {
             v.push_back(head->val);
